Unified Data Table and fixed loading issues when changing values in data table
